test(nestjs-signed-url): extract gcs client fixture
