+++ packages/nestjs-signed-url/src/gcs/tests/gateway.test.ts
@@ -1,68 +1,23 @@
-import type { GetSignedUrlConfig } from '@google-cloud/storage'
-import type { Storage }            from '@google-cloud/storage'
+import type { FakeGcsClient }   from '../../../tests/gcs.client.fixture.js'
 
-import assert                      from 'node:assert/strict'
-import { beforeEach }              from 'node:test'
-import { describe }                from 'node:test'
-import { it }                      from 'node:test'
-import { mock }                    from 'node:test'
+import assert                   from 'node:assert/strict'
+import { beforeEach }           from 'node:test'
+import { describe }             from 'node:test'
+import { it }                   from 'node:test'
+import { mock }                 from 'node:test'
 
-import { GcsSignedUrlGateway }     from '../gateway.js'
-
-interface FakeGcsFile {
-  params?: GetSignedUrlConfig
-  getSignedUrl: (params: GetSignedUrlConfig) => Promise<[string]>
-}
-
-interface FakeGcsBucket {
-  file: (filename: string) => FakeGcsFile
-}
-
-interface FakeGcsClient {
-  bucketName?: string
-  filename?: string
-  fileObject: FakeGcsFile
-  bucket: (bucketName: string) => FakeGcsBucket
-}
-
-const createFakeGcsClient = (): FakeGcsClient => {
-  const fileObject: FakeGcsFile = {
-    async getSignedUrl(params: GetSignedUrlConfig): Promise<[string]> {
-      fileObject.params = params
-
-      return ['signed-url']
-    },
-  }
-
-  const client: FakeGcsClient = {
-    fileObject,
-
-    bucket(bucketName: string): FakeGcsBucket {
-      client.bucketName = bucketName
-
-      return {
-        file(filename: string): FakeGcsFile {
-          client.filename = filename
-
-          return fileObject
-        },
-      }
-    },
-  }
-
-  return client
-}
-
-const createGateway = (client: FakeGcsClient): GcsSignedUrlGateway =>
-  new GcsSignedUrlGateway(client as unknown as Storage)
+import { GcsSignedUrlGateway }  from '../gateway.js'
+import { createFakeGcsStorage } from '../../../tests/gcs.client.fixture.js'
 
 describe('GcsSignedUrlGateway', () => {
   let gateway: GcsSignedUrlGateway
   let client: FakeGcsClient
 
   beforeEach(() => {
-    client = createFakeGcsClient()
-    gateway = createGateway(client)
+    const fixture = createFakeGcsStorage()
+
+    client = fixture.client
+    gateway = new GcsSignedUrlGateway(fixture.storage)
   })
 
   describe('generateWriteUrl', () => {
